# Overhaul dead code elimination, make --save-metas the default