--- a/test/Succeed/Issue7058/ModParamA.agda
+++ b/test/Succeed/Issue7058/ModParamA.agda
@@ -1,0 +1,5 @@
+
+-- A.agda
+module Issue7058.ModParamA where
+private P = Set₁
+module M (x : P) where