--- a/test/Succeed/Issue7058/ModParamB.agda
+++ b/test/Succeed/Issue7058/ModParamB.agda
@@ -1,0 +1,5 @@
+
+-- B.agda
+module Issue7058.ModParamB where
+open import Issue7058.ModParamA
+module M' = M Set