--- a/test/Succeed/Issue7058/RewriteB.agda
+++ b/test/Succeed/Issue7058/RewriteB.agda
@@ -1,0 +1,11 @@
+
+-- B.agda
+{-# OPTIONS --rewriting #-}
+
+module Issue7058.RewriteB where
+open import Issue7058.RewriteA
+
+postulate c : Set
+
+_ : a
+_ = c