- Fix #7058 by rewriting most of dead code elimination.
- Make `--save-metas` the default.

Dead code elimination now does the following:
- `eliminateDeadCode` computes and returns collections of reachable definitions and metas. It does not instantiate anything and has no side effect. The roots for reachability are the following:
   - Public definitions and pattern synonyms
   - Definitions with COMPILE pragma
   - Primitive definitions
   - Parameter sections of public modules
   - Local builtins
   - All rewrite rules 
- In `buildInterface`, we `instantiateFull` the output interface if we have `--no-save-metas`. 

Moreover, `--save-metas` is now made default. Reasons for doing so:
- Future optimizations of meta instantiations critically depend on `--save-metas`, and `--no-save-metas` should eventually become deprecated in favor of finer control over meta inlining.
- Hence, `--save-metas` should be the config that's being continuously tested. While making this branch, I discovered that `make test` actually failed with `--save-metas` because of an unrelated bug in `NamesIn` (which is fixed here).
- Benchmarks on my machine (AMD 7700X, DDR5 6000):
  - std-lib: checking time 5.4% shorter (204.27 -> 192.8 s), interface size +1.83% (168.9M -> 172M).
  - cubical lib checking time 0.7% longer (622.600 -> 627.2 s), interface size +4.6% (153M -> 160M).

In short there's a significant performance win in `std-lib` and an insignificant loss in `cubical`. I'd consider the interface size diffs to be also insignificant.

Right now, part of the small impact of `--save-metas` is that all definition bodies are `instantiateFull`-ed by elaboration in `Agda.TypeChecking.Rules.Def`. So far, I think that the following metas can survive elaboration: those in function clause LHS-es, rewrite rules, display forms, module parameter sections, and some metas in definitions bodies, those that become solved later by elaborating other things in a mutual group.