@plt-amy 2024-05-21
You can do [`env AGDA=/what/ever AGDAFLAGS='whatever' bash support/check.sh`](https://github.com/plt-amy/1lab/blob/main/support/check.sh)