@AndrasKovacs 2024-05-21
More or less as expected, 1lab OOM-s on `1lab.HIT.Truncation`, as described in #7231.