@plt-amy 2024-05-22
To summarize the discussion there/our findings from the AIM, the issue is essentially that keeping around the telescopes for solved (primarily instance) metas in macros leads to quadratic duplication of intermediate results:

```agda
foo : {{ _ : Class }} → TC ⊤

test : TC ⊤
test = do
  x ← something -- imagine it produces a very big result
  foo {{ ?inst ... x }}
```

After evaluating `something` and `bindTC`, we're looking at `foo {{ ?inst ... [very big result ] }}`, and this result is further duplicated in every bind that happens after `something`. I think this doesn't *usually* matter, but if something has to traverse the term (like `addContext`, for the escape check), it's explosive.