@AndrasKovacs 2024-05-22
@plt-amy 

- In #7231 there's a link for instantiated and non-instantiated terms [here](https://gist.github.com/ncfavier/b0e6aba03e4b415ef6695ea0df8684e0); neither term looks like it should cause OOM or hanging. Do we know that the issue is just non-instantiated term size there, or some actual `freeIn` bug? 
- Is it true that `TC`-produced terms are generally not instantiated, unlike ordinary definitions?