@jespercockx 2024-07-05
@AndrasKovacs What is the status of this PR? AFAICT it looks like it's ready to merge, though perhaps we should wait until after the release of 2.7 as to avoid unintended side-effects.