@AndrasKovacs 2024-07-05
I think it should be merged. The `--save-metas` is the preferred forward-looking default as I mentioned, but if something blows up it's extremely easy to just change the default back.