@jespercockx 2024-07-07
Merging this now, if there's any complaints we can always change the default back before the release.