@andreasabel 2024-07-09
@jespercockx wrote:
> Merging this now, if there's any complaints we can always change the default back before the release.

Well, this could be the cause of the fresh regression #7354.
- #7354 

Other complaints:
- no changelog
- wrong milestone (says 2.8.0)