@jespercockx 2024-07-09
Ah I'm sorry, I should've checked the CHANGELOG (@AndrasKovacs could you please add it?). 

The regression - while regrettable - is exactly the kind of valuable information we only get after merging.