@AndrasKovacs 2024-07-09
Added to CHANGELOG in #7356